move all image files to images folder
