+++ src/Pages/Confirm/Confirm.js
@@ -12,10 +12,10 @@ import "firebase/firestore";
 import { getEventsData } from "../../Utils/firebase.js";
 import { useHistory } from "react-router-dom";
 // import Swal from "sweetalert2";
-import printer from "../../Components/printer.png";
+import printer from "../../images/printer.png";
 // import link from "../../Components/link.png";
 // import shareTo from "../../Components/share.png";
-import map from "../../Components/map.png";
+import map from "../../images/map.png";
 import SocialMediaShare from "./SocialMediaShare.js";
 // import ReactDOM from "react-dom";
 import Joyride from "react-joyride";
