+++ src/images/map.png
