+++ src/images/next.png
